% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testPackageFunction.R
\name{testPackageFunction}
\alias{testPackageFunction}
\title{This is the test function title}
\usage{
testPackageFunction()
}
\arguments{
\item{x}{This is a decription of what parameter x would be if it were used in the function.}
}
\value{
This would be a description of the returned value (which should be 16).
}
\description{
This is the description of thwat the test function will do.
}
\examples{
testPackageFunction()
}
